remove api_key mixin
